third_party/blink/web_tests/platform/generic/http/tests/websocket/bufferedAmount-after-send-expected.txt

WebSocket bufferedAmount after send

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

onopen
PASS ws.bufferedAmount is 5
PASS ws.bufferedAmount is 37
PASS ws.bufferedAmount is 42
PASS ws.bufferedAmount is 45
onclose
PASS ws.bufferedAmount is 0
PASS successfullyParsed is true

TEST COMPLETE
